Initial stem trie, kinda works but needs tons of cleaning, probably rewriting some parts, renaming the class etc.
Current implementation uses a sorted array list for storing children on nodes, does a linear search if number of children is smaller than 6, or binary search otherwise. This is not the most efficient way, but I will probably use a similar technique for flat version so for now lets see how this will work. (We can later try dumb hashmap, a size(alphabet) array, a bitfied etc. )

Mechanism is a bit hard to follow, so I will clean that part up as well.

git-svn-id: http://zemberek3.googlecode.com/svn/trunk@199 e79309c6-3b30-11de-a81c-fde7d73ceb89